プレビュー機能を追加
